refactor AuthAccount type to a singleton
