+++ runtime/sema/errors.go
@@ -2398,7 +2398,7 @@ type InvalidTransactionPrepareParameterTypeError struct {
 func (e *InvalidTransactionPrepareParameterTypeError) Error() string {
 	return fmt.Sprintf(
 		"prepare parameter must be of type `%s`, not `%s`",
-		&AuthAccountType{},
+		AuthAccountType,
 		e.Type.QualifiedString(),
 	)
 }
