+++ runtime/sema/type.go
@@ -168,7 +168,7 @@ type ContainedType interface {
 type ContainerType interface {
 	Type
 	isContainerType() bool
-	NestedTypes() *StringTypeOrderedMap
+	GetNestedTypes() *StringTypeOrderedMap
 }
 
 func VisitThisAndNested(t Type, visit func(ty Type)) {
@@ -179,7 +179,7 @@ func VisitThisAndNested(t Type, visit func(ty Type)) {
 		return
 	}
 
-	containerType.NestedTypes().Foreach(func(_ string, nestedType Type) {
+	containerType.GetNestedTypes().Foreach(func(_ string, nestedType Type) {
 		VisitThisAndNested(nestedType, visit)
 	})
 }
@@ -4046,7 +4046,7 @@ func init() {
 		CharacterType,
 		&StringType{},
 		&AddressType{},
-		&AuthAccountType{},
+		AuthAccountType,
 		&PublicAccountType{},
 		PathType,
 		StoragePathType,
@@ -4588,7 +4588,7 @@ func (*CompositeType) isContainerType() bool {
 	return true
 }
 
-func (t *CompositeType) NestedTypes() *StringTypeOrderedMap {
+func (t *CompositeType) GetNestedTypes() *StringTypeOrderedMap {
 	return t.nestedTypes
 }
 
@@ -4624,613 +4624,6 @@ func (t *CompositeType) initializeMemberResolvers() {
 
 		t.memberResolvers = withBuiltinMembers(t, members)
 	})
-}
-
-// AuthAccountType represents the authorized access to an account.
-// Access to an AuthAccount means having full access to its storage, public keys, and code.
-// Only signed transactions can get the AuthAccount for an account.
-
-type AuthAccountType struct{}
-
-func (*AuthAccountType) IsType() {}
-
-func (*AuthAccountType) String() string {
-	return "AuthAccount"
-}
-
-func (*AuthAccountType) QualifiedString() string {
-	return "AuthAccount"
-}
-
-func (*AuthAccountType) ID() TypeID {
-	return "AuthAccount"
-}
-
-func (*AuthAccountType) Equal(other Type) bool {
-	_, ok := other.(*AuthAccountType)
-	return ok
-}
-
-func (*AuthAccountType) IsResourceType() bool {
-	return false
-}
-
-func (*AuthAccountType) IsInvalidType() bool {
-	return false
-}
-
-func (*AuthAccountType) IsStorable(_ map[*Member]bool) bool {
-	return false
-}
-
-func (*AuthAccountType) IsExternallyReturnable(_ map[*Member]bool) bool {
-	return false
-}
-
-func (*AuthAccountType) IsEquatable() bool {
-	return false
-}
-
-func (*AuthAccountType) TypeAnnotationState() TypeAnnotationState {
-	return TypeAnnotationStateValid
-}
-
-func (t *AuthAccountType) RewriteWithRestrictedTypes() (result Type, rewritten bool) {
-	return t, false
-}
-
-var authAccountTypeAddPublicKeyFunctionType = &FunctionType{
-	Parameters: []*Parameter{
-		{
-			Label:      ArgumentLabelNotRequired,
-			Identifier: "key",
-			TypeAnnotation: NewTypeAnnotation(
-				&VariableSizedType{
-					Type: &UInt8Type{},
-				},
-			),
-		},
-	},
-	ReturnTypeAnnotation: NewTypeAnnotation(
-		VoidType,
-	),
-}
-
-const authAccountTypeAddPublicKeyFunctionDocString = `
-Adds the given byte representation of a public key to the account's keys
-`
-
-var authAccountTypeRemovePublicKeyFunctionType = &FunctionType{
-	Parameters: []*Parameter{
-		{
-			Label:      ArgumentLabelNotRequired,
-			Identifier: "index",
-			TypeAnnotation: NewTypeAnnotation(
-				&IntType{},
-			),
-		},
-	},
-	ReturnTypeAnnotation: NewTypeAnnotation(
-		VoidType,
-	),
-}
-
-const authAccountTypeRemovePublicKeyFunctionDocString = `
-Removes the public key at the given index from the account's keys
-`
-
-var authAccountTypeSaveFunctionType = func() *FunctionType {
-
-	typeParameter := &TypeParameter{
-		Name:      "T",
-		TypeBound: StorableType,
-	}
-
-	return &FunctionType{
-		TypeParameters: []*TypeParameter{
-			typeParameter,
-		},
-		Parameters: []*Parameter{
-			{
-				Label:      ArgumentLabelNotRequired,
-				Identifier: "value",
-				TypeAnnotation: NewTypeAnnotation(
-					&GenericType{
-						TypeParameter: typeParameter,
-					},
-				),
-			},
-			{
-				Label:          "to",
-				Identifier:     "path",
-				TypeAnnotation: NewTypeAnnotation(StoragePathType),
-			},
-		},
-		ReturnTypeAnnotation: NewTypeAnnotation(VoidType),
-	}
-}()
-
-const authAccountTypeSaveFunctionDocString = `
-Saves the given object into the account's storage at the given path.
-Resources are moved into storage, and structures are copied.
-
-If there is already an object stored under the given path, the program aborts.
-
-The path must be a storage path, i.e., only the domain ` + "`storage`" + ` is allowed
-`
-
-var authAccountTypeLoadFunctionType = func() *FunctionType {
-
-	typeParameter := &TypeParameter{
-		Name:      "T",
-		TypeBound: StorableType,
-	}
-
-	return &FunctionType{
-		TypeParameters: []*TypeParameter{
-			typeParameter,
-		},
-		Parameters: []*Parameter{
-			{
-				Label:          "from",
-				Identifier:     "path",
-				TypeAnnotation: NewTypeAnnotation(StoragePathType),
-			},
-		},
-		ReturnTypeAnnotation: NewTypeAnnotation(
-			&OptionalType{
-				Type: &GenericType{
-					TypeParameter: typeParameter,
-				},
-			},
-		),
-	}
-}()
-
-const authAccountTypeLoadFunctionDocString = `
-Loads an object from the account's storage which is stored under the given path, or nil if no object is stored under the given path.
-
-If there is an object stored, the stored resource or structure is moved out of storage and returned as an optional.
-
-When the function returns, the storage no longer contains an object under the given path.
-
-The given type must be a supertype of the type of the loaded object.
-If it is not, the function returns nil.
-The given type must not necessarily be exactly the same as the type of the loaded object.
-
-The path must be a storage path, i.e., only the domain ` + "`storage`" + ` is allowed
-`
-
-var authAccountTypeCopyFunctionType = func() *FunctionType {
-
-	typeParameter := &TypeParameter{
-		Name:      "T",
-		TypeBound: AnyStructType,
-	}
-
-	return &FunctionType{
-		TypeParameters: []*TypeParameter{
-			typeParameter,
-		},
-		Parameters: []*Parameter{
-			{
-				Label:          "from",
-				Identifier:     "path",
-				TypeAnnotation: NewTypeAnnotation(StoragePathType),
-			},
-		},
-		ReturnTypeAnnotation: NewTypeAnnotation(
-			&OptionalType{
-				Type: &GenericType{
-					TypeParameter: typeParameter,
-				},
-			},
-		),
-	}
-}()
-
-const authAccountTypeCopyFunctionDocString = `
-Returns a copy of a structure stored in account storage under the given path, without removing it from storage, or nil if no object is stored under the given path.
-
-If there is a structure stored, it is copied.
-The structure stays stored in storage after the function returns.
-
-The given type must be a supertype of the type of the copied structure.
-If it is not, the function returns nil.
-The given type must not necessarily be exactly the same as the type of the copied structure.
-
-The path must be a storage path, i.e., only the domain ` + "`storage`" + ` is allowed
-`
-
-var authAccountTypeBorrowFunctionType = func() *FunctionType {
-
-	typeParameter := &TypeParameter{
-		TypeBound: &ReferenceType{
-			Type: AnyType,
-		},
-		Name: "T",
-	}
-
-	return &FunctionType{
-		TypeParameters: []*TypeParameter{
-			typeParameter,
-		},
-		Parameters: []*Parameter{
-			{
-				Label:          "from",
-				Identifier:     "path",
-				TypeAnnotation: NewTypeAnnotation(StoragePathType),
-			},
-		},
-		ReturnTypeAnnotation: NewTypeAnnotation(
-			&OptionalType{
-				Type: &GenericType{
-					TypeParameter: typeParameter,
-				},
-			},
-		),
-	}
-}()
-
-const authAccountTypeBorrowFunctionDocString = `
-Returns a reference to an object in storage without removing it from storage.
-
-If no object is stored under the given path, the function returns nil.
-If there is an object stored, a reference is returned as an optional.
-
-The given type must be a reference type.
-It must be possible to create the given reference type for the borrowed object.
-If it is not, the function returns nil.
-
-The given type must not necessarily be exactly the same as the type of the borrowed object.
-
-The path must be a storage path, i.e., only the domain ` + "`storage`" + ` is allowed
-`
-
-var authAccountTypeLinkFunctionType = func() *FunctionType {
-
-	typeParameter := &TypeParameter{
-		TypeBound: &ReferenceType{
-			Type: AnyType,
-		},
-		Name: "T",
-	}
-
-	return &FunctionType{
-		TypeParameters: []*TypeParameter{
-			typeParameter,
-		},
-		Parameters: []*Parameter{
-			{
-				Label:          ArgumentLabelNotRequired,
-				Identifier:     "newCapabilityPath",
-				TypeAnnotation: NewTypeAnnotation(CapabilityPathType),
-			},
-			{
-				Identifier:     "target",
-				TypeAnnotation: NewTypeAnnotation(PathType),
-			},
-		},
-		ReturnTypeAnnotation: NewTypeAnnotation(
-			&OptionalType{
-				Type: &CapabilityType{
-					BorrowType: &GenericType{
-						TypeParameter: typeParameter,
-					},
-				},
-			},
-		),
-	}
-}()
-
-const authAccountTypeLinkFunctionDocString = `
-Creates a capability at the given public or private path which targets the given public, private, or storage path.
-The target path leads to the object that will provide the functionality defined by this capability.
-
-The given type defines how the capability can be borrowed, i.e., how the stored value can be accessed.
-
-Returns nil if a link for the given capability path already exists, or the newly created capability if not.
-
-It is not necessary for the target path to lead to a valid object; the target path could be empty, or could lead to an object which does not provide the necessary type interface:
-The link function does **not** check if the target path is valid/exists at the time the capability is created and does **not** check if the target value conforms to the given type.
-The link is latent. The target value might be stored after the link is created, and the target value might be moved out after the link has been created.
-`
-
-var authAccountTypeUnlinkFunctionType = &FunctionType{
-	Parameters: []*Parameter{
-		{
-			Label:          ArgumentLabelNotRequired,
-			Identifier:     "capabilityPath",
-			TypeAnnotation: NewTypeAnnotation(CapabilityPathType),
-		},
-	},
-	ReturnTypeAnnotation: NewTypeAnnotation(VoidType),
-}
-
-const authAccountTypeUnlinkFunctionDocString = `
-Removes the capability at the given public or private path
-`
-
-var authAccountTypeGetCapabilityFunctionType = func() *FunctionType {
-
-	typeParameter := &TypeParameter{
-		TypeBound: &ReferenceType{
-			Type: AnyType,
-		},
-		Name:     "T",
-		Optional: true,
-	}
-
-	return &FunctionType{
-		TypeParameters: []*TypeParameter{
-			typeParameter,
-		},
-		Parameters: []*Parameter{
-			{
-				Label:          ArgumentLabelNotRequired,
-				Identifier:     "capabilityPath",
-				TypeAnnotation: NewTypeAnnotation(CapabilityPathType),
-			},
-		},
-		ReturnTypeAnnotation: NewTypeAnnotation(
-			&CapabilityType{
-				BorrowType: &GenericType{
-					TypeParameter: typeParameter,
-				},
-			},
-		),
-	}
-}()
-
-var publicAccountTypeGetCapabilityFunctionType = func() *FunctionType {
-
-	typeParameter := &TypeParameter{
-		TypeBound: &ReferenceType{
-			Type: AnyType,
-		},
-		Name:     "T",
-		Optional: true,
-	}
-
-	return &FunctionType{
-		TypeParameters: []*TypeParameter{
-			typeParameter,
-		},
-		Parameters: []*Parameter{
-			{
-				Label:          ArgumentLabelNotRequired,
-				Identifier:     "capabilityPath",
-				TypeAnnotation: NewTypeAnnotation(PublicPathType),
-			},
-		},
-		ReturnTypeAnnotation: NewTypeAnnotation(
-			&CapabilityType{
-				BorrowType: &GenericType{
-					TypeParameter: typeParameter,
-				},
-			},
-		),
-	}
-}()
-
-const authAccountTypeGetCapabilityFunctionDocString = `
-Returns the capability at the given private or public path, or nil if it does not exist
-`
-
-var accountTypeGetLinkTargetFunctionType = &FunctionType{
-	Parameters: []*Parameter{
-		{
-			Label:          ArgumentLabelNotRequired,
-			Identifier:     "capabilityPath",
-			TypeAnnotation: NewTypeAnnotation(CapabilityPathType),
-		},
-	},
-	ReturnTypeAnnotation: NewTypeAnnotation(
-		&OptionalType{
-			Type: PathType,
-		},
-	),
-}
-
-const accountTypeGetLinkTargetFunctionDocString = `
-Returns the target path of the capability at the given public or private path, or nil if there exists no capability at the given path.
-`
-
-const accountTypeAddressFieldDocString = `
-The address of the account
-`
-
-const accountTypeContractsFieldDocString = `
-The contracts of the account
-`
-
-const accountTypeStorageUsedFieldDocString = `
-The current amount of storage used by the account in bytes
-`
-
-const accountTypeStorageCapacityFieldDocString = `
-The storage capacity of the account in bytes
-`
-
-func (t *AuthAccountType) GetMembers() map[string]MemberResolver {
-	return withBuiltinMembers(t, map[string]MemberResolver{
-		"address": {
-			Kind: common.DeclarationKindField,
-			Resolve: func(identifier string, _ ast.Range, _ func(error)) *Member {
-				return NewPublicConstantFieldMember(
-					t,
-					identifier,
-					&AddressType{},
-					accountTypeAddressFieldDocString,
-				)
-			},
-		},
-		"storageUsed": {
-			Kind: common.DeclarationKindField,
-			Resolve: func(identifier string, _ ast.Range, _ func(error)) *Member {
-				return NewPublicConstantFieldMember(
-					t,
-					identifier,
-					&UInt64Type{},
-					accountTypeStorageUsedFieldDocString,
-				)
-			},
-		},
-		"storageCapacity": {
-			Kind: common.DeclarationKindField,
-			Resolve: func(identifier string, _ ast.Range, _ func(error)) *Member {
-				return NewPublicConstantFieldMember(
-					t,
-					identifier,
-					&UInt64Type{},
-					accountTypeStorageCapacityFieldDocString,
-				)
-			},
-		},
-		"addPublicKey": {
-			Kind: common.DeclarationKindFunction,
-			Resolve: func(identifier string, _ ast.Range, _ func(error)) *Member {
-				return NewPublicFunctionMember(
-					t,
-					identifier,
-					authAccountTypeAddPublicKeyFunctionType,
-					authAccountTypeAddPublicKeyFunctionDocString,
-				)
-			},
-		},
-		"removePublicKey": {
-			Kind: common.DeclarationKindFunction,
-			Resolve: func(identifier string, _ ast.Range, _ func(error)) *Member {
-				return NewPublicFunctionMember(
-					t,
-					identifier,
-					authAccountTypeRemovePublicKeyFunctionType,
-					authAccountTypeRemovePublicKeyFunctionDocString,
-				)
-			},
-		},
-		"save": {
-			Kind: common.DeclarationKindFunction,
-			Resolve: func(identifier string, _ ast.Range, _ func(error)) *Member {
-				return NewPublicFunctionMember(
-					t,
-					identifier,
-					authAccountTypeSaveFunctionType,
-					authAccountTypeSaveFunctionDocString,
-				)
-			},
-		},
-		"load": {
-			Kind: common.DeclarationKindFunction,
-			Resolve: func(identifier string, _ ast.Range, _ func(error)) *Member {
-				return NewPublicFunctionMember(
-					t,
-					identifier,
-					authAccountTypeLoadFunctionType,
-					authAccountTypeLoadFunctionDocString,
-				)
-			},
-		},
-		"copy": {
-			Kind: common.DeclarationKindFunction,
-			Resolve: func(identifier string, _ ast.Range, _ func(error)) *Member {
-				return NewPublicFunctionMember(
-					t,
-					identifier,
-					authAccountTypeCopyFunctionType,
-					authAccountTypeCopyFunctionDocString,
-				)
-			},
-		},
-		"borrow": {
-			Kind: common.DeclarationKindFunction,
-			Resolve: func(identifier string, _ ast.Range, _ func(error)) *Member {
-				return NewPublicFunctionMember(
-					t,
-					identifier,
-					authAccountTypeBorrowFunctionType,
-					authAccountTypeBorrowFunctionDocString,
-				)
-			},
-		},
-		"link": {
-			Kind: common.DeclarationKindFunction,
-			Resolve: func(identifier string, _ ast.Range, _ func(error)) *Member {
-				return NewPublicFunctionMember(
-					t,
-					identifier,
-					authAccountTypeLinkFunctionType,
-					authAccountTypeLinkFunctionDocString,
-				)
-			},
-		},
-		"unlink": {
-			Kind: common.DeclarationKindFunction,
-			Resolve: func(identifier string, _ ast.Range, _ func(error)) *Member {
-				return NewPublicFunctionMember(
-					t,
-					identifier,
-					authAccountTypeUnlinkFunctionType,
-					authAccountTypeUnlinkFunctionDocString,
-				)
-			},
-		},
-		"getCapability": {
-			Kind: common.DeclarationKindFunction,
-			Resolve: func(identifier string, _ ast.Range, _ func(error)) *Member {
-				return NewPublicFunctionMember(
-					t,
-					identifier,
-					authAccountTypeGetCapabilityFunctionType,
-					authAccountTypeGetCapabilityFunctionDocString,
-				)
-			},
-		},
-		"getLinkTarget": {
-			Kind: common.DeclarationKindFunction,
-			Resolve: func(identifier string, _ ast.Range, _ func(error)) *Member {
-				return NewPublicFunctionMember(
-					t,
-					identifier,
-					accountTypeGetLinkTargetFunctionType,
-					accountTypeGetLinkTargetFunctionDocString,
-				)
-			},
-		},
-		"contracts": {
-			Kind: common.DeclarationKindField,
-			Resolve: func(identifier string, _ ast.Range, _ func(error)) *Member {
-				return NewPublicConstantFieldMember(
-					t,
-					identifier,
-					&AuthAccountContractsType{},
-					accountTypeContractsFieldDocString,
-				)
-			},
-		},
-	})
-}
-
-var authAccountTypeNestedTypes = func() *StringTypeOrderedMap {
-	nestedTypes := NewStringTypeOrderedMap()
-	nestedTypes.Set("Contracts", &AuthAccountContractsType{})
-	return nestedTypes
-}()
-
-func (*AuthAccountType) isContainerType() bool {
-	return true
-}
-
-func (*AuthAccountType) NestedTypes() *StringTypeOrderedMap {
-	return authAccountTypeNestedTypes
-}
-
-func (*AuthAccountType) Unify(_ Type, _ *TypeParameterTypeOrderedMap, _ func(err error), _ ast.Range) bool {
-	return false
-}
-
-func (t *AuthAccountType) Resolve(_ *TypeParameterTypeOrderedMap) Type {
-	return t
 }
 
 // PublicAccountType represents the publicly accessible portion of an account.
@@ -5627,7 +5020,7 @@ func (*InterfaceType) isContainerType() bool {
 	return true
 }
 
-func (t *InterfaceType) NestedTypes() *StringTypeOrderedMap {
+func (t *InterfaceType) GetNestedTypes() *StringTypeOrderedMap {
 	return t.nestedTypes
 }
